feat: Allow multiple `mix.sass() or `mix.less()` calls to output different bundles/files.
Before this change the chunk name for stylesheets was the same as whatever we had in the first `.js()` call which works fine if all you need is 1 entry point to both JS and CSS but in most apps I believe you'll have probably something like `vendor.scss` and then an `app.scss` and you'd like those to be separate. This fix allows you to have multiple calls and output different bundles.

e.g.

```
.sass('resources/assets/sass/vendor.scss', 'public/css/vendor-styles.css')
.sass('resources/assets/sass/components.scss', 'public/css/app-styles.css')
```