Add inDevelopment method
This method comes in handy when you want to check if you're working in a development environment. As production and development are the most common environments to work in, often more names like staging, testing, etc. are used to denounce specific environments. Having a specific helper to indicate the development environment can come in handy when running certain processes which are only necessary for development.