Test refactoring (#3162)
* Remove babel config file

This shouldn't've gotten comitted :D

* Refactor vue tests

The duplication has been bothering me for a while. This eliminates that almost entirely.

* Refactor assertions

* Work on ESM test compliance

* Disable autoprefixer by default in tests

It causes a non-trivial slowdown per test class. It saves nearly 25s over the whole test suite.

* Fix typescript errors

* Only show browser logs on failure

* Lazily calculate versioning globs

Now we won't glob for anything until the build has started

* Eliminate usage of globals in components (#3155)

* Disable notifications during tests

This results in a roughly ~20s speedup on a 2018 fully-speced macbook pro

* Simplify mix.combine implementation

* Upgrade to Ava 4 and full ESM for tests

* Cache node deps for quicker test setup

* Refactor to prep for parallelization

* Refactor postcss tests

* Mark tests for serial or parallel execution

* Enable parallel tests

* Increase log timeout